Was able to work through the 02 lesson for this assignment
Only difference was a slight structure difference in how the error mess-
age is output if the number entered is less than 0, does the same thing
though. Then I actually had the i = 1 and the conditional i <= num, in
my for loop but the solution had i = 0, and i < num which results in the
same thing.