Got the 03 assignment to pass all the tests
First struggle was getting the string to slpit into an array properly,
did not realize I needed quotes with no spaces to get each character its
own array. So when it would go to "reverse" the array the array only had
one element so naturally nothing would happen. Once I fixed that I tried
using the Array.toString() function on it which comma seperates all of
elements, then realized Array.join() allows parameters for seperators,
which when set to '' just joins all the elements together with no spaces
as desired