Was able to complete the 04 assignment with all tests passing
End result was to make all fallowing arguments past the first array a
rest parameter and then used a if else statement nested in a for of loop
over all the elements of the rest parameter array to tell the code to
compare whatever element is iterated on for the rest parameter to each
element in the input array and if they strictly match (for type seneitiv
ity) to splice that index one element (ie remove the element), if not
continue onto the next iteration, that way if a element of the rest par-
ameter did not match any the rest of the elements can be compared still.
Then after the for of loop finished the function then returns the value
of the modified input array.

This was one of the harder tasks for me so far....