Was able to complete the 05 assignment.
Utilized a if else statement that checks to see if the two inputs are
the numbers type first or else it returns an error message, then there
is another if else statement that checks to see what input is larger
and if both inputs are greater than zero. If one of them is not greater
than zero it also kicks out an error message. Then depending on which
input is smaller the 'i' variable is set to that input while the
condition is set to i <= larger input, where the final expression is i++
to for the logic of i starting at the smaller input and looping until i
is greater than or equal to the larger input where the expression in the
loop is just the variable sum += i. Last part of the function is just
returning the variable sum.