Was able to complete the 06 assignment with all tests passing.
first if else check looks to see if input year is divisible by 4 and if it is not divisible by 100, if both are true it returns a true statement. If false, it then does a similar check to see if the input year is divisible by both 100 and 400 and if so returns true, then the else statement just returns false.